+++ CotEditor/Sources/FormatPaneController.swift
@@ -533,7 +533,7 @@ final class FormatPaneController: NSViewController, NSMenuItemValidation, NSTabl
     private func deleteSyntaxStyle(name: String) {
         
         let alert = NSAlert()
-        alert.messageText = String(localized: "Are you sure you want to delete “\(name)” syntax style?")
+        alert.messageText = String(localized: "Are you sure you want to delete “\(name)”?")
         alert.informativeText = String(localized: "This action cannot be undone.")
         alert.addButton(withTitle: String(localized: "Cancel"))
         alert.addButton(withTitle: String(localized: "Delete"))