+++ CotEditor/Sources/Image+Status.swift
@@ -72,10 +72,10 @@ struct StatusImage_Previews: PreviewProvider {
     static var previews: some View {
         
         VStack(alignment: .leading) {
-            Label(title: { Text("none") }, icon: { Image(status: .none) })
-            Label(title: { Text("available") }, icon: { Image(status: .available) })
-            Label(title: { Text("particallyAvailable") }, icon: { Image(status: .particallyAvailable) })
-            Label(title: { Text("unavailable") }, icon: { Image(status: .unavailable) })
+            Label(title: { Text(verbatim: "none") }, icon: { Image(status: .none) })
+            Label(title: { Text(verbatim: "available") }, icon: { Image(status: .available) })
+            Label(title: { Text(verbatim: "particallyAvailable") }, icon: { Image(status: .particallyAvailable) })
+            Label(title: { Text(verbatim: "unavailable") }, icon: { Image(status: .unavailable) })
         }.padding()
     }
 }