+++ CotEditor/Sources/InconsistentLineEndingsViewController.swift
@@ -107,8 +107,10 @@ final class InconsistentLineEndingsViewController: NSViewController {
         messageField.stringValue = {
             switch self.lineEndings.count {
                 case 0:  return String(localized: "No issues found.")
-                case 1:  return String(localized: "Found a line ending other than \(self.documentLineEnding.name).")
-                default: return String(localized: "Found \(self.lineEndings.count) line endings other than \(self.documentLineEnding.name).")
+                case 1:  return String(localized: "Found a line ending other than \(self.documentLineEnding.name).",
+                comment: "%@ is a line ending type, such as LF")
+                default: return String(localized: "Found \(self.lineEndings.count) line endings other than \(self.documentLineEnding.name).",
+                comment: "%lld is the number of inconsistent line endings and %@ is a line ending type, such as LF")
             }
         }()
     }