+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change Log
 - Fix an issue that invalid style names could be registered.
 - Fix an issue in the Key Binding settings that the documents currently opened were wrongly listed in the Window menu.
 - Fix an issue that the context menu in the editor didn't contain the script menu when the root scripts folder has only folders.
+- Fix missing localization.